Update ZaxJsonParser.h
If a string starts with char -97, then that string won't have trailing quotation marks in the resulting JSON, thus can be used to result in a valid json, if the actual string is a valid json. That optional trailing flag (-97) can also be inserted in the beginning of the string when parsing to a string.